requirements.txt
click==8.0.4
Cython==0.29.32
dataclasses==0.8
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
nudepy==0.5.1
Pillow==8.4.0
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
